+++ fragment_point.py
@@ -1,0 +1,7 @@
+fragment_point = """
+#version 120
+
+void main() {
+    gl_FragColor = vec4(1, 0, 0, 1);
+}
+"""
